src/django-backend/requirements.txt
django>=5.2
psycopg2-binary
gunicorn
django-cors-headers
whitenoise
djangorestframework
django-cors-headers
pytest
pytest-django
pylint
pylint-django
requests